requirements.txt
fastapi==0.111.0
numpy==1.24.3
Pillow==9.4.0
Pillow==10.3.0
pydantic==1.10.8
Requests==2.31.0
scikit_learn==1.3.0
tensorflow==2.13.1
tensorflow_intel==2.13.1
uvicorn==0.29.0
